Resolve "Write tests for updating a rocket"

Closes #57

See merge request TI/2020-2021/s3/project-ii/projects/groep-03/server!53

+++ src/main/java/be/howest/ti/mars/logic/data/H2Repository.java
@@ -322,7 +322,7 @@ public class H2Repository implements MarsRepository {
 
             stmt.setFloat(1, weight);
             stmt.setFloat(2, volume);
-            stmt.setDouble(3, rocketId);
+            stmt.setInt(3, rocketId);
 
             stmt.executeUpdate();
